+++ ProjectEuler/ProjectEuler/Problem4.cpp
@@ -1,11 +1,11 @@
 #include "Problem.h"
 #include "ProblemHelpers.h"
 
-long long ProblemFour() {
-    long long maxPalindrome = 0;
+int ProblemFour() {
+    int maxPalindrome = 0;
     for (int i = 999; i >= 100; i--) {
         for (int j = 999; j >= 100; j--) {
-            long long number = i * j;
+            int number = i * j;
             string possiblePalindrome = convertInt(number);
             string reverse = reverseString(possiblePalindrome);
             if (possiblePalindrome.compare(reverse) == 0)
